+++ app/components/Views/MultichainAccounts/accountDetails/accountTypes/baseAccountDetails/index.tsx
