+++ app/components/Views/MultichainAccounts/accountDetails/accountTypes/baseAccountDetails/styles.ts
