+++ app/components/Views/MultichainAccounts/accountDetails/components/accountInfo/accountInfo.styles.ts
