+++ app/components/Views/MultichainAccounts/accountDetails/components/accountInfo/accountInfo.test.tsx
